Add a "hidden" config option that allows for access from other devices.
This commit adds a config option named `localbind` that defaults to 'True`. When the value is `True` is bind to localhost (127.0.0.1). When the value is False, the function `get_ip()` determines the "primary" IP address of a device and binds to it.

Current default is intended to increase security. This setting is not exposed via the CLI or Web UIs, but must be created in the `client.settings` file.