Dependency Fixes
Looks like cudatoolkit is now shipping 11.3, but pytorch has no version for this resulting in an installation of the CPU only version. This is going to lead to people unable to get their GPU running, so for now we force the recommended 11.1 version. I also don't see any harm in allowing Python 3.9 so thats now the default as well to prevent future issues.