Override the `__close` metamethod of `FILE*` objects
This enables the use of a Lua 5.4 feature where you can have Lua
auto-close a file when exiting a block. For example:

```
do
    local f <close> = kobold.get_config_file()
end
```

If we exit from the `do` block under any circumstances, including via an
error or return statement, the `FILE*` `f` automatically closes right
before that happens.