Allow input_box resizing
For when you need a larger input box you can now resize it.

The whole UI could use an overhaul to make it dynamic. But I lack the experience to do this properly as I am not a web developer. Someone else with CSS knowledge feel free to make it so the UI allows for better resizing and automatically scales the gamescreen based on what is on display so it no longer pushes elements out of view.