Easier Dropdown support for Git
Colab does not allow to define labels for the modes, so to keep things simple for a version selector United is now capitalized and Stable has been added behaving the same as leaving it empty. This does potentially break all the dev colab's, so if you forked one make sure it uses United with a capital U for maximum compatibility.