Seperate Google Drive folders to its own section
This allows Colab developers to first get the correct folder structure on drive, before placing a configuration file for the model. That way we can quickly add the settings for a model without maintaining the init settings in two different lines. Its a substitute to the common --init only and --init skip approach from before.