Reordering Settings
More settings reordering so similar settings are on the same rows now that we have more settings for the repetition penalty. Amount to generate is now top left so some muscle memory may be lost with the temp. But the settings that control AI randomness are on the same row now, and repetition related settings are next to each other as well.