Standalone Docker
KoboldAI United now has a working official docker available at henk717/koboldai:united . In the spirit of our project, this commit open sources the files used to build the docker. docker-helper.sh is in the main folder, so that it is not overwritten during updates. Instead this file is copied by the Dockerfile and should be updated trough container updates.