Fix the streaming newline bug (for real this time I promise)
After actually investigating the cause of the bug instead of duct taping
a fix on, I have produced a better fix. The previous fix caused a bug
where there a newline was *removed* where it shouldn't have been when
undoing and redoing. The reason this bug was happening in the first
place was because some "newline regulation" code was falsely detecting
the output stream as a text chunk, and attempted to remove a newline
from there instead of the actual chunk.