Add include_anote kwarg to lua_compute_context.
Add an optional keyword argument to lua_compute_context to control
whether the author's note should be included in the context.  The
default value is true, so if the include_anote kwarg is not specified
then the author's note will be included, which was the default
behavior prior to this change.

Also update the Lua API documentation to describe this kwarg.