Fix space tokens being dropped from stream preview.
When streaming preview tokens arrive, they are accumulated into
a stream_preview span.  These tokens were inserted one-by-one
into the `innerText` property of this span.  However, the behavior
of innerText is to discard trailing whitespace, which meant
that a token that was entirely composed of spaces would be
discarded.
See this link for more information on this behavior:
https://stackoverflow.com/questions/47768523/empty-spaces-are-ignored-by-the-innertext-property

I tried fixing this by switching to use `textContent`, however this
caused newlines to be discarded instead.  This change fixes the issue
by accumulating incoming tokens into a string and then assigning the
string to innerText.