Submodule support
In the near future we will be using a git submodule  to integrate the Horde Bridge. The updater needs to support downloading submodules so that our users will automatically get them installed.

In this commit all the submodules are automatically updated when you do Kobold updates, and also installed if missing.