ix loading a non-started V1 story json in UI2
There was an exception loading or importing a V1 story json into UI2 if
the story had zero actions. Do to incorrect indendation, the V1 load
process would attempt to use an uninitialized temp_story_class variable.

However, even after fixing this, it was impossible to continue playing
the saved story, because it was not possible to submit a blank action
when a story only had a prompt and zero actions. Modify the logic around
the first submission to handle a non-empty initial prompt. This allows
the player to begin playing the loaded game.