Change the maximum memory length to be based on the token budget
The maximum memory length was hardcoded to 512 tokens independent of
the overall context token budget. This was restrictive when the token
budget was increased, and excessive when the budget was decreased.

Based on feedback, this change modifies the max memory length to be 50%
of the context length. I tested with a memory consisting of the first
five paragraphs of The Three Musketeers, and the change seemed
reasonable with context lengths of 512, 1024 and 2048.

In order to minimize changes, this commit only modifies
max_memory_length; max_prompt_length and max_authornote_length are still
hardcoded as 512 tokens. We might want to make the max prompt length
flexible in the future, but since the author's note is injected into
the story, it's probably fine to restrict its length since a long
author's note will likely derail the AI.