Tidy Alphabetical Model List
I changed the model list from a 100+ line monstrosity to an alphabetized dictionary which uses 1/4th the LOC. Functionality is untouched. (Also fixed a typo in the readme.md URL)

This is intended to make it easier to see which models are currently in the colab, both for devs (in the code) and for users (in drop-down).