WI: Don't explode when user uploads image without a save
This is a band-aid on an underlying problem: there is no save directory
to put blobs like images in before the user saves the game. The way
forward is probably to have an in-memory (no disk, Colab privacy (thats
kind of an oxymoron)) folder or something. I want to expand the data
storage functionality into an api in the future so devs can seamlessly
do something like:

Data.get_file_contents("image/blahblah.png")

and they won't have to actually worry about stuff like this